Move package rwt.internal into rap namespace
Rename package org.eclipse.rwt.internal -> org.eclipse.rap.rwt.internal

The namespace org.eclipse.rwt is not owned by the RAP project and should
not be used anymore. Except for the copied platform API, all packages
should be prefixed with the name of the bundle that exports it.
With RAP 2.0, we should move the rwt package into our namespace.

This is the first part of this move.

See bug 272792: Non-conforming use of namespace org.eclipse.rwt
https://bugs.eclipse.org/bugs/show_bug.cgi?id=272792