Move package org.eclipse.rwt into rap namespace
- Rename package org.eclipse.rwt -> org.eclipse.rap.rwt including
  subpackages
- Update @since tags to 2.0 because the package change breaks binary
  compatibility

The namespace org.eclipse.rwt is not owned by the RAP project and should
not be used anymore. Except for the copied platform API, all packages
should be prefixed with the name of the bundle that exports it.
With RAP 2.0, we should move the rwt package into our namespace.

See bug 272792: Non-conforming use of namespace org.eclipse.rwt
https://bugs.eclipse.org/bugs/show_bug.cgi?id=272792