Use context classloader to load application classes
In an application container, the RWT library may reside in a shared
folder. In this case, the classloader for RWT cannot load classes from
a web application. Hence, using the RWT classloader to load an
ApplicationConfiguration fails in RWTServletContextListener.

Use the thread's context classloader if available, otherwise fall back
on the RWT classloader. The container should ensure that this context classloader is be able to load the application's classes.

Credits go to Henning Blohm

See bug 367033: RWTServletContextListener should use context class
loader to load configurator class
https://bugs.eclipse.org/bugs/show_bug.cgi?id=367033