Consolidate negated p2 requirements to org.eclipse.rap.feature
Having the negated p2 requirements in all bundles causes downstream
builds to fail with unsatisfiable dependencies. This may be an error in
the build technology (Tycho and/or p2), but by moving the negated p2
requirements to the RAP feature this issue should be solved. If it works
further clean-up is required.