Remove PhaseListenerUtil
The phase id of a PhaseEvent can be checked by comparing it to one of
the public constants, e.g. `event.getPhaseId() == PhaseId.READ_DATA`.

The only purpose of this class was to wrap those expressions in static
helper methods. I think the benefit of this indirection is questionable
and it does not justify an additional public class.