Introduce RemoteObject
RemoteObject is used to synchronize an object with its remote object.
In contrast to the existing ClientObject, it does not render operations
immediately, but keeps operations in a queue for later rendering.
This allows to keep the rendering order defined by the lifecycle.

RemoteObject is meant to replace ClientObject in the long run.