Call create in constructor, add destroyed checks
There doesn't seem to be a reason to defer the sending of the create
operation. Thus, the create operation should be rendered by the
framework when a RemoteObject is created.

Once destroy has been called on a RemoteObject, no further operations
must be sent to the client. Thus, all subsequent calls to methods that
try to change the remote state must fail.