Introduce static lifecycle adapter for remote objects
While receiving and sending messages is controlled by the RWT lifecycle,
incoming operations must be applied in the READ_DATA phase and outgoing
operations must be rendered in RENDER in order to maintain the order of
operations in the protocol.

RemoteObjectLifeCycleAdapter is an adapter between the lifecycle and the
remote objects. It
- dispatches incoming operations to RemoteObjects in readData()
- lets RemoteObjects render queued operations in render()