Introduce handler for remote operations
Introduce RemoteOperationHandler as a simple means of processing
operations from a remote object. When a handler is registered with
a RemoteObject, it will be called for all operations from its remote.

This is a provisional design to allow experiments, we may consider to
replace it with a more advanced remote mechanism.