Revise tests and documentation for service handler
Trying to register two service handlers with the same id will result in
an InvalidArgumentException now. Before this change, the second handler
would silently overwrite the first one. This will almost certainly be a
mistake, the old behavior obfuscated the source of the problem.

Update documentation of ServiceHandler, ServiceManager, and
serviceHandler extension point.