Use entrypoint extensions also for IApplication
Since application extensions (IApplication) can not be referenced from
a branding extension anymore, we supported a "path" and a "brandingId"
parameter for application extensions instead in RAP 2.0M3. This turned
out to create 2 problems:

1) Third-party bundles may contain application extensions that lead
   to an exception in RAP because they don't have a path parameter.
2) Application extensions in third-party bundles have to be modified
   in order to be made available in RAP.

Fix this with an alternative approach: Application extensions can be
made available in RAP by a separate entrypoint extension using a new
parameter "applicationId". Path and branding can be specified in the
same way as for "normal" entrypoints.

Also add an example extension to the workbench demo to be able to test
this feature.

See bug 394805: "The parameter 'path' must not be null" when running RAP
with Workbench
https://bugs.eclipse.org/bugs/show_bug.cgi?id=394805