Rename session store to UISession
To make the difference between the HttpSession and RAP's UI session more
obvious, we decided to rename the ISessionStore to UISession.

* SessionStore -> UISession
* SessionStoreListener -> UISessionListener
* SessionStoreEvent -> UISessionEvent

Also rename internal classes and variables.

RWT has a new method getUISession(). To provide a clue for developers
migrating to RAP 2.0, the interface ISessionStore remains in place as
deprecated type. For SessionStoreListener and SessionStoreEvent, there
is no such compatibility version.