Remove HttpSessionBindingListener support in UISession
For historical reasons, UISession supported attributes that implement HttpSessionBindingListener. Whenever an attribute was set, removed, or
overwritten, and the attribute value implemented this interface, the
listener methods valueBound/Unbound have been called by the UISession.

This feature does not fit in with the distinction between UISession and
HttpSession. Apparently, it had been implemented for internal use in
the workbench. These cases are replaced by UISessionListener.