Rename IApplicationStore to ApplicationContext
The new name "ApplicationContext" is not limited to storing values, it
can later also be used to obtain application-wide instances of framework
services, such as the ThemeManager or the ResourceManager. Moreover, the
ApplicationContext is related to the ServletContext, which fits well
with the new name.

  HttpSession <---> UISession
  ServletContext <---> ApplicationContext

Suffix the existing internal class ApplicationContext with "Impl" and
let it implement ApplicationContext.

Revise API documentation, leave compatibility API for IApplicationStore
in place.