Make IServiceStore deprecated
The service store is only used internally, applications or components
should not use need to use it. If an application needs to store any
data in the scope of a request, it can use the API provided by the
request object (HttpServletRequest).

Use the service store implementation directly inside the framework.