Fix request counter replication in cluster
Some containers replicate only those session attributes that have been
marked as changed using HttpSession#setAttribute().

The RequestCounter instance is kept in the HttpSession to be able to
track concurrent access from different browser tabs. Once we implement
real multi-tab support, it can be moved to the UISession.

For now, let RWTClusterSupport refresh the request counter instance in
the HttpSession as changed, as it already does with the UISession.