Eliminate RWTFactory
RWTFactory is a facade that provides access to application-scoped
instances. It's a bunch of getters that simply delegate to the
application context.

For internal use, this facade is not really needed.
Obtaining instances directly from the ApplicationContextImpl results
in slightly more verbose code, but it makes the structure more obvious.

Move static method getApplicationContext() to ContextProvider, which
already provides access to the service context and UISession.

Use static import for ContextProvider.getApplicationContext(). I think
we should generally use static imports for ContextProvider#getXxx.
This makes the code more readable and compensates for the added
verbosity. As an example, instead of

    RWTFactory.getEntryPointManager()

we have now

    getApplicationContext().getEntryPointManager()