Allow RemoteObject#set to overwrite previous value
It's legal to call set sequentially for the same property on a remote
object, e.g. when the same value changes multiple times in a row.

When no other operation is rendered meanwhile, the previous value
will be overridden, otherwise a new set operation will be created.

Fix bug 397545: Java Remote Object throws exception if "set" is called
  twice for same property
https://bugs.eclipse.org/bugs/show_bug.cgi?id=397545