Append an identifier to UISession id
UISession#getId() returns the id of the HttpSession, which once was the
purpose of this method. It should now return the id of the UISession.

Appending the string ":1" to distinguish it from the HttpSession and to 
intentionally break code that uses the method in the old way. The
HttpSession id still be obtained by uiSession.getHttpSession().getId().