Declare ILifeCycle deprecated
This interface is only used to register phase listeners. This should not
be done at runtime anymore, but only in an application configuration.

Phase listeners have application scope, which makes them unsuitable for
short-time, session-specific uses (requires checking the session display
in the listener code). In the past, we used those session-scoped phase
listeners to render custom code to the client. With the new protocol and
the JavaScriptExecutor service, this should not be needed anymore.

Phase listeners are going to be replaced by the new RemoteObject API.
New components should use this API instead of PhaseListener.