Fix: do not remove phase listener in another session's request
The code that triggers the additional layout was only effective if the
phase listener is called for the same UI session first. In the likely
case that another session's request triggers the phase listener before,
the listener will be removed without doing anything.

I couldn't come up with a test, but we'll revisit this code when we
eliminate phase listeners.