Introduce abstract base class for entrypoints
This abstract base class implements of the common tasks that are needed
by all entrypoints: it creates a display and a main shell, it opens the
shell and it runs the SWT event loop if required by the current
operation mode.

Using this base class should be preferred over implementing the
EntryPoint interface itself, because it saves some boilerplate code
and is compatible with both operation modes.