Provide additional Juno Equinox target feature
There is now a new org.eclipse.rap.equinox.juno.target.feature feature
that contains the same dependencies as the regular feature for Kepler
(org.eclipse.rap.equinox.target.feature) but it uses Juno SR1 as input
source for building. Both features are aggregated into the final
repository, but not both of them are visible in the p2 categories.