Revise JavaDoc for RWT.getRequest/Response
The framework uses requests to communicate with the client.
Applications should not need to access the request and response
objects. However there are some rare exceptions, like accessing
request parameters or setting cookies.

Declare RWT.getReqeust/getResponse as "not recommended" and explain
that application should normally not use these methods.