Replace try-catch in SessionLocaleProvider
The extension registry is also accessed outside of a UISession, e.g.
when reading global contributions at application startup. This is not
an exceptional case and should not be handled by an exception.

An odd side effect is that all these exceptions, although catched, end
up in the profiler logs.