Do not render remote objects that are already disposed
Widgets are not rendered if they are created and disposed in the same
request. Remote objects are. This behavior is not consistent and leads
to problems when a remote object refers to a widget.

The solution is to omit rendering of remote objects that are created
and disposed in the same request.

Fixes bug 395272: [ClientScripting] ClientListener and disposed Widgets
https://bugs.eclipse.org/bugs/show_bug.cgi?id=395272