Reduce the border-radius of ProgressBar to 15px
In order to correctly visualize the rounded corners of the ProgressBar
there is an adjustment for min/max indicator length (see
ProgressBar.js#_getIndicatorMin/MaxSafeLength()). It depends on the
border-radius set. The minimum value of the progress indicator is
limited to the border-radius. The original value of the border-radius
was too big (40px), which forced the min/max value of progress indicator
to 40px/width-40px too. 

402954: [ProgressBar][Client] Selection is not correctly mapped to
indicator size
https://bugs.eclipse.org/bugs/show_bug.cgi?id=402954