Add experimental support for "generic" type handler
Adding a "isGeneric" flag to the type handler will cause it to ignore
the "events", "methods" and "properties" fields and always execute all
operations. Instead of calling a specific setter, all properties are
applied at once using a "set" method (if present). Some details of this
feature are not ideal yet, like the case if there is no "set" method, or
how the method can distinguish between calls by the type handler and by
other client components.