Cut off trailing ".0" from JSON numbers
When a JsonValue number is created from double or float without a
fractional part, the created JSON should not contain the trailing ".0".

Since JSON and JavaScript do not distinguish between floating point and
integer numbers, the additional information that the number "once was"
floating point does not provide any useful information in this context.