Pick up performance improvement from JSON parser
Typical JSON messages include objects with only a few members.
Looking up object members can be improved by storing the hash of
a name in a map that points to the index in the names list.

For small item counts, an array of 32 bytes provides the best
trade-off between footprint, hashing overhead, and lookup time.