Move JSON parser into public package
JsonValue and subtypes must be public API in order to allow for
custom JSON values (such as nested JSON arrays and objects) in the
RemoteObject API.

Since the JSON parser is not related to RWT, the package is called
org.eclipse.rap.json. We may consider to move the parser into a
separate bundle at some point in the future.