Replace Map and Object in RemoteObject API with JSON types
The untyped parameters in RemoteObject and OperationHandler are
replaced with the new JSON types JsonValue and JsonObject.

With this change, the remote API is now typesafe and flexible.
The framework does not need to try to convert untyped objects,
arrays, and maps to JSON and back again anymore.

The API in org.eclipse.rap.remote id no longer marked as provisional.

RemoteObject:

  set( String, Object )
  -> ( String, JsonValue )

  call( String, Map<String, Object> )
   -> ( String, JsonObject )

OperationHandler, AbstractOperationHandler:

  handleSet( Map<String, Object> properties )
        -> ( JsonObject )

  handleCall( String, Map<String, Object> )
        -> ( String, JsonObject )

  handleNotify( String, Map<String, Object> )
        -> ( String, JsonObject )