Change test bundles to depend on mockito-all
We recently had a lot of trouble with the mockito bundle from Orbit
which has a dependency on hamcrest that conflicts with JUnit's use
of hamcrest (for details see bug 403676).

These problems do not occur with the mockito-all bundle from Maven.
This bundle also includes objenesis, so that we get rid of this extra
dependency.

* Change test bundles to Require-Bundle mockito-all 1.9.5
* Change pom to use the newly created extra-dependencies repository
  for 2.1 which includes mockito-all 1.9.5 from Maven