Allow for re-sending requests
In a case of a timeout, clients should be able to abort and send the
same request again. There can be cases when the previous request had
already been processed by the server but the response got lost on the
way back. In these cases, the server will receive the previous request
once again.

This case is now handled in LifeCycleServiceHandler by checking the
request counter. When a request is received with the previous request
counter, this request will not be processed but the previous response
will be returned.

Clients are responsible for choosing a reasonable timeout and aborting
the old request before sending it again.

See bug 383397: Supporting unreliable networks by adding a request timeout handling
https://bugs.eclipse.org/bugs/show_bug.cgi?id=383397