Replace IClientObject with RemoteObject
IClientObject was a predecessor of the RemoteObject with almost
identical methods. In contrast to IClientObject, the RemoteObject
does not offer a create method, but appends a create operation on
creation.

- Rename ClientObject to LifeCycleRemoteObject

A RemoteObject implementation for use in the lifecycle (LCAs). In
contrast to RemoteObjectImpl, this implementation writes directly to
the protocol message.

- Rename ClientObjectFactory to RemoteObjectFactory

A helper class for LCAs to create LifeCycleRemoteObjects. Besides a
getRemoteObject method, this helper now offers a createRemoteObject
method to be used instead of IClientObject.create().

- Remove IClientObjectAdapter

This class was only used for the creation of remote objects for GCs.
An instance was kept in a field in WidgetAdapterImpl, where it consumed
memory for every widget, not only Canvas. Creating IDs for GC remote
objects by adding a suffix to the widget id allows us to get rid of
this instance.