Keep all RemoteObjects in RemoteObjectRegistry
There are currently two types of RemoteObject implementations:
* DeferredRemoteObject for remote objects created by ConnectionImpl
* LifeCycleRemoteObject for remote objects used in LCAs
Both now extend the abstract base class RemoteObjectImpl, which stores
id and OperationHandler.

LifeCycleRemoteObjects have been kept only in the service store before,
but they must be kept in the session in order to retain the reference
to the operation handler.

Destroyed RemoteObjects are removed from the registry *after* render
by the RemoteObjectLifeCycleAdapter.