Send multiple display resize events in one request
See Bug 231751 - Shell#controlResize is fired too often 

With this change a request is sent no more than once every 500ms (as
opposed 60 ms before). However, the request still contains all events.
The way the server currently processes the message, only the last event
is processed, but this might change with RAP 2.2 and we should re-visit
this issue until release.