Revise ApplicationContext.add/remove listener wrt concurrency
Introduce an enum for the different lifecycle states to avoid a
confusing bookkeeping with different boolean variables. Use an
AtomicReference since the value will be accessed and modifed from
different threads.

Change ApplicationContext.add/removeListener methods to return a
boolean that indicates whether the operation was successful.
Without this return value, users of this API cannot be certain that
the listener was added/removed successfully. The context could be
destroyed just in this moment.