Add JavaDoc for row templates, minor API changes
Accept -1 (SWT.DEFAULT) for Cell width, height, and bindingIndex. Once
we introduce public getters, they will have to return -1 for undefined
values. Moreover, SWT.DEFAULT is valid for width and height in SWT
GridData as well.

For the same reasons, accept null for colors, font and image.

Throw IllegalStateException instead of IAE in case of setting one too
many dimension. It's not the argument that is invalid but the state of
the cell.

Throw NPE instead of IAE in case of an illegal null parameter.
Using NPE vs. IAE for null parameters is controversial, however NPE is
used throughout the RWT API (ParamCheck.notNull). We shouldn't change it
at this point.

Move TemplateSerializer into the correct package.