Add system property to suppress workbench autostart
The workbench bundle contains an internal ApplicationConfiguration that
is auto-started in the Activator by default. When workbench applications
require custom additions to the ApplicationConfiguration, we need to be
able to suppress this autostart behavior.

Setting the new system property "org.eclipse.rap.workbenchAutostart"
to "false" will now suppress the workbench autostart.

Applications can now subclass the WorkbenchApplicationConfiguration
and start the subclassed config manually.

* Rename WorkbenchApplicationConfigurator to
  WorkbenchApplicationConfiguration to match the implemented interface
* Make it non-final and add a default constructor to enable subclassing
  without adding a dependency to OSGi API.

See bug 377414: Allow using the workbench with a custom ApplicationConfiguration
https://bugs.eclipse.org/bugs/show_bug.cgi?id=377414