Ensure ApplicationContext is alive in beforeDestroy()
When ApplicationContextListener.beforeDestroy() is called, the
ApplicationContext must still be active in order to allow listener
code to access attributes in its cleanup procedure.

This behavior is similar to UISessionListener.beforeDestroy().

This requires that ApplicationContext.isActive() returns true while
the listeners are called, i.e. after deactivate has been called.

As a consequence, callers of deactivate should not need to check
isActive() before calling deactivate(). Instead of throwing ISEs,
activate and deactivate methods are now idempotent.

Change the contract of ApplicationRunner.start() not to throw an ISE
if called twice to be symmetric to stop().

With the previous implementation, calling ApplicationRunner.start()
twice would have removed the ApplicationContext from the ServletContext,
leaving the system in a broken state.