Ensure JRE compatibility in build, enable integration tests
Use the animal-sniffer maven plugin [1] to ensure that source code is
compatible with the Java 1.5 class library. With this change, we can
safely run our build with a JDK 6 or higher without the risk of
building code that accidentally uses a newer class library feature.

This allows us to re-enable the integration tests.

[1] http://mojo.codehaus.org/animal-sniffer-maven-plugin/usage.html

Fix bug 377497