Fix test errors in Eclipse (objenesis)
With the change from mockito-all to mockito-core in commit e8dc858e,
running the tests in Eclipse leads to NoClassDefFound errors for
classes from objenesis that are required by mockito. For some reason,
the transitive dependency from mockito-core to objenesis does not end
up on the test project's class path.

Add org.objenesis as a workaround for this problem.

Change-Id: I109847b98645845011923c47e4898dae6abb027e