Copy pristine Equinox servletbridge from plugins/ to lib/ directory
org.eclipse.equinox.servletbridge lives in two locations: One location
is the plugins directory together with all the other bundles, the other
location is the lib/ directory of the .war file. Both versions need to
be in sync. In earlier times the lib/ servletbridge jar was part of the
root files and checked into the source code repository; now it is a
unmodified copy from the plugins/ servletbridge.

424984: WAR build does not work in RAP 2.3
https://bugs.eclipse.org/bugs/show_bug.cgi?id=424984

Change-Id: I3707e69a5b3a14d7219b19503e78dc0cb9278aaa