Block non-number characters in spinner
The code responsible for blocking the non-number characters exists in
Spinner.js#_onkeypress, but is not called as "keypress" event is stopped
from propagation is BasicText.js#_onkeypress. Registering a separate
text field "keypress" listener that points to existing "_onkeypress"
handler solves the problem.

Improve blocking code to allow "-" only at the beginning.

424776: [Spinner] Allows any character
https://bugs.eclipse.org/bugs/show_bug.cgi?id=424776

Change-Id: I002d39bfbcf7406ee4931a079f1caa835ad34b07