Remove features from demo .war builds and use plugin-based products
The .war builds of the example/demo applications in
releng/org.eclipse.rap.examples.build are currently based on features.
This patch moves the bundle dependency information from the features
into the product specifications. In addition to that it moves the
rootfiles (web.xml, launch.ini) from the feature to the product; the
rootfiles are added to the WEB-INF/ directory in the assembly step, and
the features are removed completely.

Bug 425569: Remove features from WAR builds
https://bugs.eclipse.org/bugs/show_bug.cgi?id=425569

Change-Id: I0c76a4f7319c1e979d53e0b61bdb78b7452178f5