Show ErrorBox on ServerPush connection errors
When a server push request fails (connection error, e.g. because the
connection is intercepted by a proxy), the client gives up after three
retries and initiate a new UI request (POST). When this request fails,
an error dialog will be displayed to notify the user of the connection
problem.

In case of a failure that is not identified as a connection problem
(e.g. a server error), an error will now be displayed.

Some rename refactorings.

Resolves bug 422411: [webclient] Connection failure not raised when
server-push enabled
https://bugs.eclipse.org/bugs/show_bug.cgi?id=422411

Change-Id: Iea78beeb2a9be6c3ca883cd6fb5616a68be76826