Remove faulty template item content render optimization
This caused items to not be re-rendered when their content changed. This
optimization was in place to prevent image flickering issues in google
chrome. To compensate, Style.js now "remembers" if an image was already
set as an elements background image and in that case does not do so
again.

Also added some jsDoc to Template.js

See Bug 425011 - [Template] Tree have problems with filters

Change-Id: Ib80665375186269cd60b955bcaebd33aba480e56