Enable entrypoints at the root path
It should be possible to register the main entrypoint of an application
at the root path in order to be available directly at the application's
base URL.

The root path poses a special case in the Servlet API. A servlet must
be mapped to "" to be available at the root path ("/" stands for the
"default" servlet). Hence, request.getServletPath() returns "", not "/".
This behavior must be reproduced in rwt.osgi CutOffContextPathWrapper.
We must ensure that our codes handles the case that getServletPath()
returns "".

Moreover, for a request to a root servlet, the contextPath is also
empty even though the application is bound to a context path!
Therefore, contextPath and servletPath do not add up to the requestURI
anymore. When constructing URLs, we must use request.getRequestURI()
instead of request.getContextPath() + request.getServletPath().

A request to the base URL will usually end with a slash ("myapp/").
In this case, the path info is "/" but must not be redirected to the
same URL without trailing slash.

Remove unused utility class URLHelper.

Change-Id: I1b382986d2d199074455afe24b7898317414229b