Simplify SingletonManager by using AtomicReference
Using AtomicReference as container type for the singleton instances
allows to simplify the SingletonManager. A singletonHolder contains
the singleton instance for a given type and acts as a type-specific
lock at the same time.

This two-step implementation ensures that the SingletonManager itself
is not locked during the creation of a singleton instance.

1) When a singleton is requested, we first obtain the unique instance
of the singleton holder for this type using the SharedInstanceBuffer.
This operation is synchronized in SharedInstanceBuffer. However, the
creation of this AtomicReference is cheap.

2) Then we check-and-create the instance in another synchronized block
which locks only this specific singleton holder.

This implementation is simpler and faster because it requires only a
single hash lookup instead of two.

Create a test util that allows us to add the attribute store behavior
to mocked UISessions.

Change-Id: I4ea09211cef2decafbfd0b653bf36572e85346b9