Preserve the actual text selection value
When a verify listener is attached to the Text the selection value is
preserved manually - see TextLCAUtil#readTextAndSelection.
When client sends invalid selection (selectionStart is bigger than text
length) the server adjusted the value in Text#setSelection(). To avoid
rendering of this adjusted value back, the LCA should preserve the
actual value and not the one sent by the client.

424118: [Text] Cursor reset to first character when typing fast
https://bugs.eclipse.org/bugs/show_bug.cgi?id=424118

Change-Id: If8c72e196b3b64b6ad1038f900e1651c90c04d4d