Preserve the actual text selection value
When a verify listener is attached to Text/Combo/CCombo the selection
value is preserved in ProcessActionRunner - see
TextOperationHandler#handleSetSelection for example.
When client sends invalid selection (selectionStart is bigger than text
length - see bug 427015) the server adjusts the value in
Text#setSelection(). To avoid rendering of this adjusted value back to
the client, the we should preserve the actual value (like regular
preserving does) and not the one sent by the client.

424118: [Text] Cursor reset to first character when typing fast
https://bugs.eclipse.org/bugs/show_bug.cgi?id=424118

Change-Id: I40e44b096966017a96848cb5ff9d584d09e4b999