Introduce SingletonUtil methods with scope parameter
Accepting the scope as a parameter allows to access singletons from
background threads without running in a "fake context". This pattern
also allows for better unit tests.

Adding two "getUniqueInstance" methods to SingletonUtil, one for
UISession and one for ApplicationContext. This name reveals the purpose
and the function of this method even when static imports are used.

A possible alternative would be to add getSingletonInstance(type)
methods to UISession and ApplicationContext. This had resulted in a
clearer API, but it had put the debatable concept of singletons at a
very prominent place in the RAP API. Support for scoped singletons is
really just a utility and should remain decoupled.

Replace ContextProvider.getUISession with RWT.getUISession in the old
method SingletonUtil.getSessionInstancethis to enable inline method
refactoring. Whether the old method SingletonUtil.getSessionInstance
should be deprecated is up to another discussion.

As an alternative implementation, the SingletonManager could have been
kept in a field in ApplicationContextImpl. I decided against that
because it's not necessary for an ApplicationContext to work.

Bug 425573: Add SingletonUtil method with UISession parameter
https://bugs.eclipse.org/bugs/show_bug.cgi?id=425573

Change-Id: I3278d8a031b6726ea64d639bddcacc65e72dc61c