Decouple clientbuilder from rwt
Removing the dependency to rwt from the clientbuilder allows us to use
it for other projects as well. Moreover, this is a precondition for
using the client builder at build time - when the rwt build depends on
the clientbuilder, we'd otherwise end up with a circular dependency.

Read the list of JavaScript files to be included in the client.js from
a text file `client.files`.

Accept this file as a command line parameter to JsCompressor. Also add
a parameter for the base path that includes these files.

Move ClientResources from displaykit to internal.resources package.

Change-Id: I2672aec7a08ee5d3732768019835b10124eca153