Create client.js in Maven build, exclude from git repository
Creating the compressed JavaScript library file is actually a build
step. Doing this in the Maven build ensures that it cannot be forgotten.
It is still possible to create a client.js manually for testing.

Exclude the client.js from the git repository as it is not a build
artifact. Since this file changes with every JavaScript change,
including it in the git repository prevented git from merging
otherwise unrelated commits in the past.

Change-Id: I42b8383898db73ed6a255ced75edf7a1734411f4