Basic support for old JavaScript unit tests in Jasmine
This is work in progress and does not remove functionality from either
the existing TestRunner or the Jasmine runner. Currently the Jasmine
runner requires an additional parameter to run the old style tests,
e.g.: "jasmine.html?legacy=true&filter=Grid" (filter is optional)

Changes to the unit test classes themselves are either to make them work
with the new implementation of the asserts (which are mapped to the
Jasmine matchers that may behave differently in some cases), or to clean
up the DOM (which wasn't as much of an issue before since the old
TestRunner was able to cover everything else, Jasmine somehow isn't.)

Remaining issues: 
- Not all tests are green yet, and not all clean up the DOM 100%.
- Tests with asynchronous calls are currently skipped. We will have to
see if we automatically convert them also, or manually convert them to
real Jasmine Specs right away.
- While Jasmine is capable to run only specific Specs/Suites, it needs
way more time to skip the other tests than is acceptable. That's why the
old-style "filter" attribute is still supported, but we will need a
solution for pure Jasmine Suites also.
- Test with very old browser.
- Remove the old TestRunner and clean up the code. 

Change-Id: I97771462989469b8207e9c66e1070dffb3102c9d