Allow mouse events outside client area
Mouse events behavior in SWT (Windows) is the following:
- not fired when click on control border
- not fired when click on Shell titlebar and menubar
- not fired when click on Tree/Table header for Tree/Table mouse
listener
- not fired when click on Scrollbars for control mouse listener
- fired when click on Group label, TabFolder/CTabFolder items bar

In long run, we should fix this entirely on the client to not send mouse
events in these specific areas. For Tree/Table this will be tricky,
because the same client-side widget is used in Nebula Grid
implementation, where mouse events are fired in column header area. 

Change ControlOperationHandler to reject only mouse events with negative
coordinates (click on border).
Change ShellOperationHandler to reject mouse events in titlebar and
menubar.
Change Tree/TableOperationHandler to reject mouse events in header.
Stop the prapagation of mousedown/mouseup events in ScrollBar.js
Adjust client and server tests accordingly.

394624: MouseClick on composite not always dispatched
https://bugs.eclipse.org/bugs/show_bug.cgi?id=394624

Change-Id: Idd5149c931595f58b601de668fe70f6d17b8633e
Signed-off-by: Ivan Furnadjiev <ivan@eclipsesource.com>