Add client-side support for cross-application file Drag & Drop
This makes the following additions to the protocol:
- DropTarget has a new property "fileDropEnabled" (boolean)
- Drop-accept events have a new optional property "files" (the content
is an array of file-ids, but may have to be changed to include names)
- A new "service" object FileUploader with one methdod "submit" that
takes an arary of file-ids and an URL

The Server currently ignores these changes.

The EventHandler.js has been adjusted to prevent any action when a file
is dropped outside of a DropTarget.

Change-Id: Ia52f63d117ca365a5831f8e395951ffb7cb6b6d6
Signed-off-by: Tim Buschtöns <tbuschto@eclipsesource.com>