Remove broken redirect for trailing slash in URL
The RWTServlet used to redirect requests with a trailing slash to the
URL without the trailing slash, e.g. for the servlet path /foo, it
would redirect /foo/ to /foo. This redirect feature has been broken
with the introduction of root path support ("/" as a servlet path).

Since /foo/ is an invalid URL when the servlet path is /foo, I think
that a 404 "Not Found" is even more appropriate. After all, if we
redirected /foo/ to /foo, why don't we redirect /foo/bar to /foo ?

See bug 429041: Redirect loop when URL ends with /
https://bugs.eclipse.org/bugs/show_bug.cgi?id=429041

Change-Id: Ic7709f2a00fc4f467a86c26a6fb30b866ac9ab77
Signed-off-by: Ralf Sternberg <rsternberg@eclipsesource.com>