Mark ProcessActionRunner as deprecated
We're going to deprecate the lifecycle package in this release in order
to replace the concept of lifecycle phases in RAP 3.0.

Move ProcessActionRunner implementation into an internal package and
delegate to this package from the deprecated public class. This allows
us to use the internal class in our LCAs and avoid a lot of warnings.

Change-Id: I16fd014f7cd8adcad5e9da57bdf1541e74b9d2e6
Signed-off-by: Ralf Sternberg <rsternberg@eclipsesource.com>