Mark WidgetLCAUtil as deprecated
We're going to deprecate the lifecycle package in this release in order
to replace the concepts of lifecycle, phases, and LCAs in RAP 3.0.

Move WidgetLCAUtil implementation into an internal package and keep the
public class as a delegator to the internal one. This allows us to use
the internal class in our LCAs and avoid a lot of warnings.

Create a copy of the test for the implementation, but keep the tests for
the API methods in the public package to ensure the deprecated API still
works as expected.

Change-Id: I8d4039506ff992e80f7bcf674eed1d39cd9c82df
Signed-off-by: Ralf Sternberg <rsternberg@eclipsesource.com>