Mark AbstractWidgetLCA as deprecated
We're going to deprecate the lifecycle package in this release in order
to replace the concepts of lifecycle, phases, and LCAs in RAP 3.0.

Move AbstractWidgetLCA implementation into an internal package and let
the public class inherit from it. This allows us to use the internal
class in our LCAs and avoid a lot of warnings.

Create a copy of the test for the implementation, but keep the tests for
the API methods in the public package to ensure the deprecated API still
works as expected.

Change-Id: If5119157c667fd3bbfb6223f3305090937bed935
Signed-off-by: Ralf Sternberg <rsternberg@eclipsesource.com>