Redirect requests to root servlet without trailing slash
When an entrypoint is registered at the root path, and the application
is deployed using the servlet bridge, the RWTServlet can be accessed
with the contextPath with or without a trailing slash. If the trailing
slash is missing, all relative links to resources are broken [1].

In both cases (with and without trailing slash), pathInfo is "/" and
servletPath is "". However, the RequestURL contains the trailing slash
if it's present.

If the trailing slash is missing, redirect to the URL with the trailing
slash included (HTTP 302). This is the correct base URL of the web app,
also the Tomcat manager points to base URLs with trailing slash.

Responding with a 404 would also be an option, but since most servlet
containers would automatically redirect to the correct URL, this seems
to be the expected behavior.

[1] See bug 429443: Requests to Root Path Entry Point are not redirected
correctly in WAR deployment
https://bugs.eclipse.org/bugs/show_bug.cgi?id=429443

Change-Id: Ida1295f7dfe41c3856509b3ab0ff04927bfb693e
Signed-off-by: Ralf Sternberg <rsternberg@eclipsesource.com>