Prepare deprecation of WidgetUtil
In our effort to deprecate the lifecycle package, we need to start using
an internal WidgetUtil because the getLCA method of the public class now
returns an incompatible type.

At the moment, only WidgetUtil.getLCA is marked as deprecated, for the
other methods we need to provide alternatives first.

Move WidgetUtil implementation into an internal package and keep the
public class as a delegator to the internal one. This allows us to use
the internal class in our LCAs and avoid a lot of warnings.

The public method getLCA() is a special case: it has to provide an
instance of the public AbstractWidgetLCA even when the widget uses an
interal AWLCA. Since the instance cannot be casted, return a wrapper
that extends from the public class and delegates to the internal LCA.
With this implementation, this method always returns a new instance.
However, since LCAs are stateless, this should be acceptable.

Create a copy of the test for the implementation, but keep the tests for
the API methods in the public package to ensure the deprecated API still
works as expected.

Change-Id: I51871b5a2421bff88e4f5d33432761bab7ebbaec
Signed-off-by: Ralf Sternberg <rsternberg@eclipsesource.com>