Fix images not appearing in grid rows that were cleared
See Bug 429217 - [Tree] Image of TreeItem disappears after scrolling.

The issue was caused by NOT using rwt.html.Style.setBackgroundImage to
clear the images when "null" was rendered as an item. The Style util
stores additional information on the HTML element that are not updated
in this case. The bug was caused
by commit aaa36da014099d86e507de345b2726e35d9cb75e.

Change-Id: Ibf636f7a16fc176ee2f80fb41ed68a9d00460fd2
Signed-off-by: Tim Buschtöns <tbuschto@eclipsesource.com>