Mark WidgetAdapter as deprecated
We're going to deprecate the lifecycle package in this release in order
to replace the concept of lifecycle phases in RAP 3.0.

Move WidgetAdapter interface to an internal package and let the public
interface inherit from it. This allows us to use the internal interface
in our code and avoid deprecation warnings.

Change-Id: Id34fba6b248bd3dc352c5140dd5bafed523e8021
Signed-off-by: Ralf Sternberg <rsternberg@eclipsesource.com>