Mark WidgetLifeCycleAdapter as deprecated
We're going to deprecate the lifecycle package in this release in order
to replace the concept of lifecycle phases in RAP 3.0.

Move WidgetLifeCycleAdapter interface to an internal package and let the
public interface inherit from it. This allows us to use the internal
interface in our code and avoid deprecation warnings.

Change-Id: I417b32c8b246b50a622b39ee58720916152f4610
Signed-off-by: Ralf Sternberg <rsternberg@eclipsesource.com>