Mark PhaseListeners as deprecated
In our effort to process protocol operations in order, we need to
replace the phase model. Thus the concept of phases must be removed
from our API.

In contrast to other deprecated types in the lifecycle package, the
typesafe enum pattern in PhaseId makes it impossible to delegate to
to an internal copy of the class. Therefore, the types are simply
deprecated and will be moved to the internal package in 3.0.

* Mark these types as deprecated: PhaseListener, PhaseEvent, PhaseId
* Suppress deprecation warnings in RWT

Change-Id: I642dbdbff5cb58b65b0cc242456ec47dff506b6c
Signed-off-by: Ralf Sternberg <rsternberg@eclipsesource.com>